[publish][m]: fix the validation for non tabular data
We use goodtable for validate tabular data. In the previous implementation
we used to pass datapackage.json as an argument to goodtables.Inspector.
The problem was that, if any non tabular data present in datapackage it was
also trying to validate that. That is what we could not publish any non tabular data.

In this commit, tried to iterate through all data and validate one by one if it is tabular
 data. We just simply skip the validation if the data is non tabular.

 We try to determine if the data is tabular or not from the descriptor format, mediatype or simply
 checking the file ends with csv or not.

 Added one fixture geojson data for test and fixed two old tests as in this implementation the
 report format has been changed.

Fixes : #89